requirements.txt
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
djangorestframework==3.16.1
elastic-transport==8.17.1
elasticsearch==8.19.2
idna==3.11
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.6.1
